Move extra functions to separate header file
